+++ docs/conf.py
@@ -48,16 +48,16 @@ master_doc = 'index'
 
 # General information about the project.
 project = u'HydrOffice Sound Speed'
-copyright = u'2025, CCOM/JHC,UNH'
+copyright = u'2026, CCOM/JHC,UNH'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
 # built documents.
 #
 # The short X.Y version.
-version = '2025.5'
+version = '2026.1'
 # The full version, including alpha/beta/rc tags.
-release = '2025.5.0'
+release = '2026.1.0'
 
 numfig = True